requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-telegram-bot[job-queue]==21.5
python-dotenv==1.0.1
groq>=0.9.0
